# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API for saving films you want to watch.",
        "title": "Watchlist API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/collections": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of collections by user ID from authentication token. It also returns metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get user collections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by `name`",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Specify the desired `page`",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Specify the desired `page size`",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting by `id`, `name`, `created_at`. Use `-` for desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Add a new collection. You will be granted the permissions to get, update, and delete it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Add new collection",
                "parameters": [
                    {
                        "description": "Information about the new collection",
                        "name": "collection",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collection_id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get the collection by ID. You must have permissions to get this collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update the collection by ID. You must have the permissions to update it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Update the collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New information about the collection",
                        "name": "film",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.CollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.FilmResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Delete the collection by ID. You must have the permissions to delete it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Delete the collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "filters.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "description": "The current page number.",
                    "type": "integer",
                    "example": 2
                },
                "first_page": {
                    "description": "The first page number, usually 1.",
                    "type": "integer",
                    "example": 1
                },
                "last_page": {
                    "description": "The last page number based on total records and page size.",
                    "type": "integer",
                    "example": 3
                },
                "page_size": {
                    "description": "The number of items per page.",
                    "type": "integer",
                    "example": 5
                },
                "total_records": {
                    "description": "The total number of records available.",
                    "type": "integer",
                    "example": 15
                }
            }
        },
...
        "models.Collection": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "Timestamp when the collection was created.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "description": {
                    "description": "Description of the collection; optional, up to 500 characters.",
                    "type": "string",
                    "maxLength": 500,
                    "example": "This is description"
                },
                "id": {
                    "description": "Unique identifier for the collection.",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name of the collection; required, between 3 and 100 characters.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "My collection"
                },
                "updated_at": {
                    "description": "Timestamp when the collection was last updated.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "user_id": {
                    "description": "Identifier of the user who created the collection.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.Film": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "comment": {
                    "description": "URL of the film's photo; optional, must be a valid URL.",
                    "type": "string",
                    "maxLength": 500,
                    "example": "This is comment"
                },
                "created_at": {
                    "description": "Timestamp when the film was added.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "description": {
                    "description": "Description of the film; optional, up to 1000 characters.",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is description"
                },
                "genre": {
                    "description": "Genre of the film; optional, only alphabetic characters.",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Horror"
                },
                "id": {
                    "description": "Unique identifier for the film.",
                    "type": "integer",
                    "example": 1
                },
                "is_viewed": {
                    "description": "Indicates if the user has viewed the film.",
                    "type": "boolean",
                    "example": false
                },
                "photo_url": {
                    "description": "URL of the film's photo; optional, must be a valid URL.",
                    "type": "string",
                    "example": "https://placeimg.com/640/480"
                },
                "rating": {
                    "description": "Rating of the film; optional, must be between 1 and 10.",
                    "type": "number",
                    "maximum": 10,
                    "minimum": 1,
                    "example": 6.7
                },
                "review": {
                    "description": "User's review of the film; optional, up to 500 characters.",
                    "type": "string",
                    "maxLength": 500,
                    "example": "This is review"
                },
                "title": {
                    "description": "Title of the film; required, between 3 and 100 characters.",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "My film"
                },
                "updated_at": {
                    "description": "Timestamp when the film details were last updated.",
                    "type": "string",
                    "example": "2024-09-04T13:37:24.87653+05:00"
                },
                "user_id": {
                    "description": "Identifier of the user who added the film.",
                    "type": "integer",
                    "example": 1
                },
                "user_rating": {
                    "description": "User's rating of the film; optional, between 1 and 10.",
                    "type": "number",
                    "maximum": 10,
                    "minimum": 1,
                    "example": 5.5
                },
                "year": {
                    "description": "Release year of the film; optional, must be between 1888 and 2100.",
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 1888,
                    "example": 2001
                }
            }
        },
...
        "swagger.CollectionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is description"
                },
                "name": {
                    "type": "string",
                    "example": "My collection"
                }
            }
        },
        "swagger.CollectionResponse": {
            "type": "object",
            "properties": {
                "collection": {
                    "$ref": "#/definitions/models.Collection"
                }
            }
        },
        "swagger.CollectionsResponse": {
            "type": "object",
            "properties": {
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Collection"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/filters.Metadata"
                }
            }
        },
        "swagger.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some kind of error"
                }
            }
        },
...
        "swagger.FilmResponse": {
            "type": "object",
            "properties": {
                "film": {
                    "$ref": "#/definitions/models.Film"
                }
            }
        },
...
        "swagger.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "some kind of success message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}